trust_align/requirements.txt
absl-py==2.1.0
accelerate==0.34.0
aiohttp==3.9.5
aiosignal==1.3.1
git+https://github.com/huggingface/alignment-handbook.git@27f7dbf00663dab66ad7334afb7a1311fa251f41
annotated-types==0.6.0
anthropic==0.26.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.5.1
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
cmake==3.30.0
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses-json==0.6.5
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.14.4
defusedxml==0.7.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dm-tree==0.1.8
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
e==1.4.5
einops==0.8.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
etils==1.7.0
evaluate==0.4.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.115.0
fastjsonschema==2.19.1
filelock==3.13.1
FlagEmbedding==1.2.9
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
fuzzywuzzy==0.18.0
gguf==0.10.0
git-filter-repo==2.38.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
hf_transfer==0.1.6
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.0
interegular==0.3.3
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.1.20
langchain-anthropic==0.1.13
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.7
langchain-text-splitters==0.0.1
langcodes==3.4.0
langsmith==0.1.54
language_data==1.2.0
lark==1.1.9
Levenshtein==0.26.0
lightgbm==4.3.0
llvmlite==0.43.0
lm-format-enforcer==0.10.3
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mauve-text==0.3.0
mdurl==0.1.2
mistral_common==1.4.3
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nmslib==2.1.1
notebook==7.1.3
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
onnxruntime==1.17.3
openai==1.50.0
openpyxl==3.1.5
orjson==3.10.3
outlines==0.0.46
overrides==7.7.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.11.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.1
plotly==5.22.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyairports==2.1.1
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybind11==2.6.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjnius==1.6.1
pynndescent==0.5.13
pynvml==11.5.0
pyparsing==3.1.2
pyserini==0.35.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-Levenshtein==0.26.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
RapidFuzz==3.10.0
ray==2.31.0
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tokenizer==3.4.3
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.45.1
triton==3.0.0
trl==0.11.1
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.4
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vllm==0.5.4
vllm-flash-attn==2.6.1
wandb==0.17.0
wasabi==1.1.2
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.16.0
xformers==0.0.27.post2
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
flash-attn==2.6.3